move plot_model() into its own module
