+++ pdr_backend/aimodel/test/test_plot_model.py
@@ -1,0 +1,1 @@
+# nothing here: put all the testing in test_aimodel_factory.py
